backend/requirements.txt
Flask
psycopg2-binary
python-dotenv
requests
Cerberus